Put examples folder in Mongo
