+++ libs/databases/mongo/demo/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "demo"
+name = "mongodb_demo"
 version = "0.1.0"
 edition = "2021"
