+++ libs/databases/mongo/demo/src/main.rs
@@ -1,5 +1,3 @@
-mod t1_async_api;
-
 fn main() {
-    let _ = t1_async_api::main();
+	println!("MongoDB demo...");
 }
